setup friendlist and add prefetch

+++ frontEnd/src/Pages/HomePage.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { connectSocket, socket } from "../Utils/socket";
-import { useQuery } from "@tanstack/react-query";
+import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { useLocation } from "react-router-dom";
 import {
   ChatRecentContainer,
@@ -14,42 +14,47 @@ import {
   HomePageMessageContainer,
   HomePageTopContainer,
 } from "../Utils/Styles/Home.style";
-import { getAllChats } from "../Utils/chat.api";
-import { useLocalStorage } from "../Utils/hooks/useLocalStorage";
+import { getAllChats, getAllPastMessage } from "../Utils/chat.api";
+import { useLocalStorage } from "../hooks/useLocalStorage";
 import { getAllUsers } from "../Utils/user.api";
 
 const HomePage = () => {
   const [msgNotification, setMsgNotification] = useState([]);
-  const useStorage = new useLocalStorage();
   const navigate = useNavigate();
   const params = useLocation();
+  const queryClient = useQueryClient();
 
   const userQuery = useQuery(["Users"], getAllUsers);
 
   const chatQuery = useQuery({
     queryKey: ["Chat"],
-    queryFn: () => getAllChats(useStorage.getUserId()),
+    queryFn: () => getAllChats(useLocalStorage.getUserId()),
     onSuccess: (data) => {
-      connectSocket(useStorage.getUser());
+      connectSocket(useLocalStorage.getUser());
     },
   });
 
-  const startChat = async (id) => {
-    // await queryClient.prefetchQuery({
-    //   queryKey: ["Private_Chat", id],
-    //   queryFn: () => getAllPastMessage(useStorage.getUserId(), id),
-    // });
+  const startChat = async (userId ,id) => {
+    await queryClient.prefetchQuery({
+      queryKey: ["Private_Chat", id],
+      queryFn: () => getAllPastMessage(userId, id),
+    });
     navigate(`/chat/?${id}`);
   };
 
   useEffect(() => {
-    socket.on("notification", ({ from, newMessage }) => {
+    socket.on("notification", ({ from, to, newMessage }) => {
       const isAlreadyDisplay = msgNotification.filter((msg) => {
         return msg.id == from;
       });
 
-      if (isAlreadyDisplay.length === 0 && params.pathname === "/") {
+      if (
+        isAlreadyDisplay.length === 0 &&
+        params.pathname === "/" &&
+        to === useLocalStorage.getUserId()
+      ) {
         setMsgNotification([...msgNotification, { id: from, newMessage }]);
+        
         localStorage.setItem(
           "notification",
           JSON.stringify({ from, newMessage })
@@ -62,7 +67,7 @@ const HomePage = () => {
     <HomePageContainer>
       <HomePageTopContainer>
         <h3>Welcome, back!</h3>
-        <strong>{useStorage.getUser().username}</strong>
+        <strong>{useLocalStorage.getUser().username}</strong>
         <HomePageFriendContainer>
           <h3>New Sneeks</h3>
 
@@ -73,7 +78,7 @@ const HomePage = () => {
                 return (
                   <div key={user._id}>
                     <HomePageFriendPicContainer
-                      onClick={() => startChat(user._id)}
+                      onClick={() => startChat(useLocalStorage.getUserId(),user._id)}
                     ></HomePageFriendPicContainer>
                     <p>{user.username}</p>
                     <div
@@ -89,11 +94,11 @@ const HomePage = () => {
       <HomePageMessageContainer>
         <h3>Recent Chat</h3>
         {chatQuery?.data?.data.messages.map((chat) => {
-          const { message, sender, time } = chat;
+          const { message, sender, time, users } = chat;
           return (
             <ChatRecentContainer
               key={chat._id}
-              onClick={() => startChat(sender)}
+              onClick={() => startChat(users[0], users[1] )}
             >
               <ChatRecentPic></ChatRecentPic>
               <ChatRecentContent>
